WIP -- add XoSmartContract
Gets rid of the original XO main and stores it in a new file main-old,
mostly useful at this point to keep around to base the work off of.
Basically main-old is unnecessary.

Also adds an initial pass at the XoSmartContract and SmartContract impl
in the /handler module of simple_xo.

Signed-off-by: Shannyn Telander <telander@bitwise.io>