Change Signer to HashSigner in sawtooth unit test
Changes the signature of the `create_batch` function in the sawtooth
unit test module to take a HashSigner reference rather than a Signer
reference. `Signer` is a trait which would require using 'dyn' in front
of it, which results in errors about this argument not having known size
at compile time. These errors can be avoided by using HashSigner
instead, which is an implementation of the Signer trait.

Signed-off-by: Shannyn Telander <telander@bitwise.io>